Move search into own component
New sub search components define results/filters and do fetching.
